Refactor verify components
Moved Verify page to elements and use better name to verification pending form
